Add files via upload
It's a multiclss Neural Network classifcation. Which the data can be any number of features. while goes to the task it's a classification. Algorthim is that Feed forward neural network. And the learning algorthim is gradient decent  with backpropgation in vectorised way. while comes to loss function is cross entropy. Last but least step is to evaluate ur model which i use that Top k accuracy score.